Moved both pickers to seperate component
