+++ lib/screens/created_reward.dart
@@ -7,6 +7,7 @@ import 'package:habitflow/blocs/rewards_bloc.dart';
 import 'package:habitflow/components/color_picker.dart';
 import 'package:habitflow/components/neu_card.dart';
 import 'package:habitflow/components/neu_text_field.dart';
+import 'package:habitflow/components/pickers.dart';
 
 /// A screen which allows user to create a reward.
 class CreateReward extends StatefulWidget {
@@ -24,17 +25,12 @@ class _CreateRewardState extends State<CreateReward> {
   Icon _icon = const Icon(Icons.accessibility);
   Color _color = Colors.redAccent;
 
-  /// Changes [_color] to one selected by user.
-  void _onColorChange(Color color) => setState(() => _color = color);
-
-  /// Allows to pick an icon.
-  Future<void> _pickIcon() async {
-    final IconData icon = await FlutterIconPicker.showIconPicker(
-      context,
-      iconPackMode: IconPack.material,
-    );
-    _icon = Icon(icon);
-    setState(() {});
+  /// Changes [_color] and [_icon] to what user selected.
+  void _onPick(Color color, Icon icon) {
+    setState(() {
+      _color = color;
+      _icon = icon;
+    });
   }
 
   /// Creates the reward.
@@ -74,61 +70,7 @@ class _CreateRewardState extends State<CreateReward> {
                   key: _formKey,
                   child: Column(
                     children: <Widget>[
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceEvenly,
-                        children: <Widget>[
-                          Column(
-                            children: <Widget>[
-                              const Padding(
-                                padding: EdgeInsets.all(8.0),
-                                child: Text(
-                                  'Icon',
-                                  style: TextStyle(
-                                    fontSize: 16.0,
-                                    fontWeight: FontWeight.w600,
-                                  ),
-                                ),
-                              ),
-                              NeuCard(
-                                depth: 1.5,
-                                radius: 100.0,
-                                child: IconButton(
-                                  onPressed: _pickIcon,
-                                  color: _color,
-                                  icon: _icon != null
-                                      ? Icon(_icon.icon)
-                                      : const Icon(Icons.accessibility),
-                                ),
-                              ),
-                            ],
-                          ),
-                          Column(
-                            children: <Widget>[
-                              const Padding(
-                                padding: EdgeInsets.all(8.0),
-                                child: Text(
-                                  'Color',
-                                  style: TextStyle(
-                                    fontSize: 16.0,
-                                    fontWeight: FontWeight.w600,
-                                  ),
-                                ),
-                              ),
-                              NeuCard(
-                                depth: 1.5,
-                                radius: 100.0,
-                                child: Padding(
-                                  padding: const EdgeInsets.all(10.0),
-                                  child: ColorPickerButton(
-                                    _onColorChange,
-                                    _color,
-                                  ),
-                                ),
-                              ),
-                            ],
-                          ),
-                        ],
-                      ),
+                      Pickers(_color, _icon, _onPick),
                       const Padding(
                         padding: EdgeInsets.all(16.0),
                         child: Divider(),
